+++ Semana 1/react/public/script.js
@@ -1,0 +1,9 @@
+/**
+ * Babel: Conventer (transpilar - converter um codigo em outro) codigo do react para um que o browser entenda.
+ * Webpack: Para cada tipo de arquivo(.js,.css,.png) eu vou converter o codigo de uma maneira diferente
+ * 
+ * Loader: Utilizados pelo webpack para que o JS entenda
+ * Babel-loader : Converte o JS para algo que o 
+ * Image-loader : Converte o JS para algo que o 
+ * Css-loader : Converte o JS para algo que o 
+ */