+++ Semana 1/react/src/index.js
@@ -1,0 +1,6 @@
+import {soma} from './soma'
+
+console.log(soma(3, 5));
+console.log(soma(6, 4));
+console.log(soma(9, 25));
+console.log(soma(91, 25));