+++ Semana 1/react/src/soma.js
@@ -1,0 +1,8 @@
+const soma = (a, b) => {
+    return a + b;
+  };
+
+
+module.exports ={
+    soma
+}