+++ Changelog.md
@@ -1,3 +1,27 @@
+# [0.0.4] - 22/08/2020
+## Semana 1 folder inprovement
+## React project creation
+## Webpack configuration
+## Babel configuration
+
+**Changed File**
+
+        modified:   Changelog.md
+        renamed:    Semana 1/package.json -> Semana 1/node/package.json
+        renamed:    Semana 1/src/index.js -> Semana 1/node/src/index.js
+        renamed:    Semana 1/yarn.lock -> Semana 1/node/yarn.lock
+        new file:   Semana 1/react/babel.config.js
+        new file:   Semana 1/react/package.json
+        new file:   Semana 1/react/public/bundle.js
+        new file:   Semana 1/react/public/index.html
+        new file:   Semana 1/react/public/script.js
+        new file:   Semana 1/react/src/index.js
+        new file:   Semana 1/react/src/soma.js
+        new file:   Semana 1/react/webpack.config.js
+        new file:   Semana 1/react/yarn.lock
+
+---
+
 # [0.0.3] - 21/08/2020
 ## Concept of Middleware
 ## Improving routes